solder les compte 6 et 7 repreise du code
