+++ plan-comptable.js
@@ -1,0 +1,33 @@
+export const PLAN_COMPTABLE = {
+    "106000": ["réserves"],
+    "119000": ["report à nouveau (solde débiteur)"],
+    "120000": ["résultat de l'exercice (excédent)", "excédent"],
+    "129000": ["résultat de l'exercice (déficit)", "déficit"],
+    "275000": ["dépôts et cautionnements versés", "cautions"],
+    "370000": ["stocks de marchandises", "inventaire"],
+    "404000": ["fournisseurs d'immobilisations"],
+    "467000": ["autres comptes débiteurs ou créditeurs", "remboursements", "prêts"],
+    "512000": ["banques"],
+    "530000": ["caisse"],
+    "602600": ["emballages"],
+    "603000": ["variations des stocks"],
+    "604000": ["achats d'études et prestations de services", "achats prestations"],
+    "606000": ["achats non stockés de matière et fournitures", "fournitures", "décorations", "énergies"],
+    "607000": ["achats de marchandises", "marchandises"],
+    "613000": ["locations"],
+    "616000": ["primes d'assurances", "assurances"],
+    "618300": ["documentation technique", "documentations"],
+    "618500": ["frais de colloques, séminaires, conférences", "conférences"],
+    "622000": ["rémunérations d'intermédiaires et honoraires", "intermédiaires"],
+    "623000": ["publicité, publications, relations publiques", "communication"],
+    "624100": ["transports sur achats"],
+    "625000": ["déplacements, missions et réceptions", "déplacements", "restauration", "hébergements"],
+    "626000": ["frais postaux et de télécommunications", "internet", "frais postaux", "télécommunications", "domiciliations"],
+    "627000": ["services bancaires et assimilés", "services bancaires"],
+    "706000": ["prestations de services", "formations"],
+    "707000": ["ventes de marchandises", "ventes"],
+    "754100": ["dons manuels", "dons"],
+    "756000": ["cotisations"],
+    "890000": ["bilan d'ouverture"],
+    "891000": ["bilan de clôture"]
+};
